Organized, Sliced & Revamped for Bugfixing
All Assets have more or less been reorganised, and the hierarchy got an overhaul (allowing for easier editing)

Everything that is not being used has been deleted, to help improve performance in Unity.

The Bathroom should be a lot easier to work with in code now, and a lot of visual bugs with lighting / doors have also been fixed (thanks to paint.net)!

Most importantly, each area now has it's own items and spot in the Hierarchy (Top Floor, Bottom Floor, Outdoors), which is def worth checking out.

With greetings and Merry Christmas! -Benjamin