Final Item Update + Changing Bools to Static
Along with the former update, this update is about Item Overhauling - making sure the items work as they should!

TV & Radio are now interactable, and will play responding SFX sounds & music.

This is also with the addition of making static bools that load across scenes, so that if you leave your TV on or your radio on, it will be still be turned on until you turn it off and vice versa.

It looks a lot cleaner and performs a lot smoother, because you don't have to fetch gameObject player, along with a lot of components anymore!