Player & Cat now 95% Working!
Player and Cat Animations now have Idle & Walk Animations that work really well - Christopher from 2nd Year helped solve it with me! : )

The last 5% is just tweaking audio with the cat & player when it comes to footsteps, but that's for later when game is close to being in a good finished state.