Radio, Lamp, TV, Cat, etc. + Other Tweaks
Fixes to the above - they're working pretty solid now in accordance with the Datatransfer Script where static bools contain information across scenes if they are active or not!

A bit hard to work with in Editor (you'll have to go to Datatransfer and set the script to true), but it functions well in testing. : )

If there are issues, these objects are at least 90% there.