FloorSwaps & Triggerbox Fixes 90%!
Swapping between TopFloor & BottomFloor should work well now, both in Editor and PlayMode without having to manually enable/disable gameObjects!

This should work independently of whether you are starting as the player on the TopFloor, or BottomFloor.

All that remains is to fix the BathroomTrigger & the bed, before going over to SceneLoad to be able to move between scenes : )