...Final Major Update to Triggers & Floors?
All floors and transitions between triggers should be okay now!

Collisions have been improved too, and a lot of shadows as well + tweaks with VFX and the glassDoor animation.

The Bedroom Door is now interactable, and can be used to open or close it.

Overall game is very ready for SceneController stuff and adding in the Item Interactions, before Timeline and Events : )