Fixed the audio sliders + radio!
The audio sliders now work as intended (had to match the volume number in Inspector on the sliders, to be -80 - 20, as it is set in AudioMixer). :speaker:

The Radio also works well after refactoring, but I'm planning 2 minor tweaks regarding the audio when radio is turned off, and figuring out if there's a problem if people spam interact on the radio. :radio: