Big Bugfix Update + tidying and cleaning code
Like the title says, this is a big Update, where a lot of bugs have been dealt with, and functionality is as it should be.

Interact-able Items now have an easier remade script for just loading optional timelines when you click yes, all that matters is to just create these timelines.

What remains is the 2nd floor, and 1 issue regarding the cat, but good progress is being made!

(I can't remember all, but at least 10+ bugs have been squashed today) :sunglasses: