Made music play on TopFloor & Outdoors
Like the title says, but I did this by moving and creating an object that is set on the spot where the Radio is. This Object called BackgroundMusic now plays all the audio, and the radio with sprite itself is just receiving inputs from player when to mute or unmute.

The radio automatically checks when you're on different floors to adjust the volume.