Overhaul to triggers, and fixes to StairsTrigger
Turned out the script and object called "ChangeBools" actually caused a few issues with how it was running in Update and enforcing the game to always be on bottomfloor, and cat visible even on TopFloor.

I fixed it, but generally did some cleaning up in code and tweaks, which should help make things run smooth.