Final bugs fixed? Wooo!
As it says in the title, I've managed to fix the remaining bugs!

I think most of it should be gone, and all that remains is to create timelines that are executed when "yes" on interactable items; to create cutscenes, and transitions to scenes (next day, or death etc).

If there is time, it would also be nice to create a shader that grants a blue outline around items when you can interact with them. : )

...There might be one more push later on to help fix any forgotten bugs / implement features, possibly with lighting or a certain item or two (e.g. radio, cat food, fridge).

There's also plans to create a script (bed?) for checking if the house is locked on Scene 3, when transitioning to Scene 4 - if not locked, then cat dies.